requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
bleak==0.22.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
fastapi==0.115.12
ffmpeg-python==0.2.0
flatbuffers==25.2.10
fonttools==4.57.0
future==1.0.0
google-auth==2.38.0
google-genai==1.9.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
msgpack==1.1.0
numba==0.61.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
sockets==1.0.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.1
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webrtcvad==2.0.10
websockets==15.0.1
zeroconf==0.146.3